% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_select.r
\name{opt_bw}
\alias{opt_bw}
\title{Bandwidth Selector}
\usage{
opt_bw(X, Y, cutpoint = NULL, verbose = FALSE, kernel = "triangular")
}
\arguments{
\item{X}{Nx1 vector containing running variable}

\item{Y}{Nx1 vector containing response variable}

\item{cutpoint}{Location of discontinuity.}

\item{verbose}{Logical indicator whether to print more information to terminal. Default is FALSE.}

\item{kernel}{Sting indicating which kernel to use. Options are "triangular", "rectangular", "epanechnikov", "quartic", "triweight", "tricube", "gaussian", and "cosine".}
}
\value{
Return optimal bandwidth
}
\description{
opt_bw calculates the Imbens-Kalyanaraman optimal bandwidth for local linear regression for RDD. This is a data-driven, asymptotically optimal choice of smoothing parameter. Conditions for asymptotic optimality (smoothness and iid X,Y) and description of the algorithm can be found in the referenced paper.
}
\examples{
X <- runif(1000, -1, 1);
Y <- 10 + 5*X + 3*(X>=0)+rnorm(1000);
opt_bw(X,Y)
}
\references{
Imbens, Guido and Karthik Kalyanaraman. (2009) "Optimal Bandwidth Choice for the regression discontinuity estimator," NBER Working Paper Series. 14726.
}
\author{
Samuel Gailliot, Texas A (and) M University, College Station, TX. samuel.gailliot@stat.tamu.edu
}
