% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_plot.r
\name{rd_plot}
\alias{rd_plot}
\title{rd_plot}
\usage{
rd_plot(
  Y,
  X,
  cutpoint = 0,
  bin_method = "esmv",
  kernel_method = "uni",
  ci = NULL,
  title = NULL,
  x.lab = NULL,
  y.lab = NULL,
  x.lim = NULL,
  y.lim = NULL
)
}
\arguments{
\item{Y}{Nx1 vector containing dependent variable}

\item{X}{Nx1 vector containing running variable (forcing variable)}

\item{cutpoint}{Location of discontinuity.}

\item{bin_method}{specifies the procedure to select the number of bins. Options are:

\code{es}: IMSE-optimal evenly-spaced method using spacings estimators.

\code{espr}: IMSE-optimal evenly-spaced method using polynomial regression.

\code{esmv}: mimicking variance evenly-spaced method using spacings estimators. Default

\code{esmvpr}: mimiching variance evenly-spaced method using  polynomial regression

\code{qs}: IMSE-optimal quantile-spaced method using spacings estimators.

\code{qspr}: IMSE-optimal quantile-spaced method using polynomial regression.

\code{qsmv}: mimicking variance quantile-spaced method using spacings estimators.

\code{qsmvpr}: mimicking variance quantile-spaced method using polynomial regression.}

\item{kernel_method}{specifies the vernel function used to construct local-polynomial estimators. Options are: \code{triangular}, \code{epanechnikov}, and \code{uniform}}

\item{ci}{optional graphical option to display confidence intervals of selected level (ex. 95) for each bin.}

\item{title}{optional title for the plot}

\item{x.lab}{optional title for the x-axis}

\item{y.lab}{optional title for the y-axis}

\item{x.lim}{optional setting for the range of the x-axis}

\item{y.lim}{optional setting for the range of the y-axis}
}
\value{
\code{rdplot} a standard \code{ggplot} object that can be used for further customization.
}
\description{
rd_plot calls the powerful function rdplot from rdrobust but makes only the most important options available to the user to make the function easier to use.
}
\examples{
X <- runif(1000, -1, 1)
Y <- 10 + 5*X + 3*(X>=0)+rnorm(1000)
rd_plot(Y,X)
}
\references{
Calonico, S., M. D. Cattaneo, M. H. Farrell, and R. Titiunik. 2017. rdrobust: Software for Regression Discontinuity Designs. Stata Journal 17(2): 372-404.

Calonico, S., M. D. Cattaneo, and R. Titiunik. 2014. Robust Data-Driven Inference in the Regression-Discontinuity Design. Stata Journal 14(4): 909-946.

Calonico, S., M. D. Cattaneo, and R. Titiunik. 2015a. Optimal Data-Driven Regression Discontinuity Plots. Journal of the American Statistical Association 110(512): 1753-1769.

Calonico, S., M. D. Cattaneo, and R. Titiunik. 2015b. rdrobust: An R Package for Robust Nonparametric Inference in Regression-Discontinuity Designs. R Journal 7(1): 38-51.

Cattaneo, M. D., B. Frandsen, and R. Titiunik. 2015. Randomization Inference in the Regression Discontinuity Design: An Application to the Study of Party Advantages in the U.S. Senate. Journal of Causal Inference 3(1): 1-24.
}
\author{
Samuel Gailliot, Texas A (and) M University, College Station, TX. samuel.gailliot@stat.tamu.edu
}
